+++ DungeonSlime/Content/Content.mgcb
@@ -31,6 +31,13 @@
 /processorParam:Quality=Best
 /build:audio/theme.ogg
 
+#begin fonts/04B_30_5x.spritefont
+/importer:FontDescriptionImporter
+/processor:FontDescriptionProcessor
+/processorParam:PremultiplyAlpha=True
+/processorParam:TextureFormat=Compressed
+/build:fonts/04B_30_5x.spritefont
+
 #begin fonts/04B_30.spritefont
 /importer:FontDescriptionImporter
 /processor:FontDescriptionProcessor
