+++ SaintGameLibrary/Scenes/Scene.cs
@@ -1,0 +1,102 @@
+﻿using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework.Content;
+using System;
+
+namespace SaintGameLibrary.Scenes;
+
+public abstract class Scene : IDisposable
+{
+    /// <summary>
+    /// Creates a new scene instance.
+    /// </summary>
+    public Scene()
+    {
+        Content = new ContentManager(Core.Content.ServiceProvider)
+        {
+            RootDirectory = Core.Content.RootDirectory
+        };
+    }
+
+    ~Scene() => Dispose(false);
+
+
+    /// <summary>
+    /// Gets the ContentManager used for loading scene-specific assets.
+    /// </summary>
+    /// <remarks>
+    /// Assets loaded through this ContentManager will be automatically unloaded when this scene ends.
+    /// </remarks>
+    protected ContentManager Content { get; }
+
+    /// <summary>
+    /// Gets a value that indicates if the scene has been disposed of.
+    /// </summary>
+    public bool IsDisposed { get; private set; }
+
+    /// <summary>
+    /// Initializes the scene.
+    /// </summary>
+    /// <remarks>
+    /// When overriding this in a derived class, ensure that base.Initialize()
+    /// still called as this is when LoadContent is called.
+    /// </remarks>
+    public virtual void Initialize()
+    {
+        LoadContent();
+    }
+
+    /// <summary>
+    /// Override to provide logic to load content for the scene.
+    /// </summary>
+    public virtual void LoadContent() { }
+
+    /// <summary>
+    /// Unloads scene-specific content.
+    /// </summary>
+    public virtual void UnloadContent()
+    {
+        Content.Unload();
+    }
+
+    /// <summary>
+    /// Updates this scene.
+    /// </summary>
+    /// <param name="gameTime">A snapshot of the timing values for the current frame.</param>
+    public virtual void Update(GameTime gameTime) { }
+
+    /// <summary>
+    /// Draws this scene.
+    /// </summary>
+    /// <param name="gameTime">A snapshot of the timing values for the current frame.</param>
+    public virtual void Draw(GameTime gameTime) { }
+
+    /// <summary>
+    /// Disposes of this scene.
+    /// </summary>
+    public void Dispose()
+    {
+        Dispose(true);
+        GC.SuppressFinalize(this);
+    }
+
+    /// <summary>
+    /// Disposes of this scene.
+    /// </summary>
+    /// <param name="disposing">'
+    /// Indicates whether managed resources should be disposed.  This value is only true when called from the main
+    /// Dispose method.  When called from the finalizer, this will be false.
+    /// </param>
+    protected virtual void Dispose(bool disposing)
+    {
+        if (IsDisposed)
+        {
+            return;
+        }
+
+        if (disposing)
+        {
+            UnloadContent();
+            Content.Dispose();
+        }
+    }
+}
